backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
aiofiles==24.1.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
python-dotenv==1.0.1
redis==5.2.1
psycopg2-binary==2.9.9
websockets==15.0.1
requests==2.32.3

# Logging and monitoring
structlog==24.4.0
loguru==0.7.2
prometheus-client==0.21.0

# Additional dependencies
celery==5.4.0
flower==2.0.1
slowapi==0.1.9

# AI Model APIs
openai==1.58.1
anthropic==0.40.0
google-generativeai==0.8.3